Proper support for sessions/windows/panes
Now it's possible to target arbitrary sessions/windows/panes. The
current pane is used as the target with the default invocation.

It's possible to read from one pane and output to another pane. To do
this, give :pipe-pane the parameters of another, valid combination.

Please see the README.md for more details.

Signed-off-by: Slinky <slinky@iki.fi>