Improve tmux and shell expansions avoidance
For some reason, tmux display-message expands e.g. ".%30" into a string
of length 30 with the "." at the beginning and "%30" at the end. This
messes up displaying the pane-ids. Fix it by double quoting the %%. Only
the tmux display-message is impacted by this.

When sending keys with tmux send-keys, make sure to send whatever we are
given as-is. Don't let the shell expand variables or otherwise modify
the string.

Signed-off-by: Slinky <slinky@iki.fi>