Separate the customizable pexpect script
Separate the user-modifiable parts (the pexpect script) from the main
echelon.py. This way the echelon.py can be more easily updated
separately from the script. Also, it's easier to have several scripts.

This change also means that from now on, your scripts should go to the
script.py file, and not the main echelon.py.

Signed-off-by: Slinky <slinky@iki.fi>